Start working on tests for ConfigService
This required the creation of FileSystemService. I still need to work on some things for that, like how the interfaces are defined. I realized I actually should be defining the interfaces in a more granular way, as needed by the consumer(s). This way, if different consumers need different subsets of functionality, test doubles only need to have those subsets instead of the whole set. It should be up to the consumer(s) to define the interface specifying what it needs.